requirements.txt
pytest
python-dotenv
requests
langchain-community
langchain-ollama
langchain_ollama

faiss-cpu
numpy

reportlab
nltk

pymupdf

sentence-transformers


PyPDF2

faker

pdfplumber

